Update README (more info in the description)
I have finally decided to test the Smart Card module. Apple says macOS supports USB SC readers. The one in the laptop got detected on Windows, so I checked it's BIOS path. Turns out it was on HS10 USB port, so I added that port to USBPorts.kext in the EFI. The device is now recognized as a Smart Card reader on macOS, and I'm pretty sure it works! I'm gonna get a regular Smart Card to test this, because it might be a pretty useful feature.

I've also checked whether CtlnaAHCIPort kext is required. On my unit I couldn't see the AHCI controller if it was disabled, so I updated the README.